resetting config to default config
